--- index.js
@@ -1,7 +1,0 @@
-exports.Connection = require('./lib/connection.js').Connection
-exports.Queue = require('./lib/queue.js').Queue
-exports.Worker = require('./lib/worker.js').Worker
-exports.MultiWorker = require('./lib/multiWorker.js').MultiWorker
-exports.Scheduler = require('./lib/scheduler.js').Scheduler
-exports.PluginRunner = require('./lib/pluginRunner.js')
-exports.Plugin = require('./lib/plugin.js').Plugin
